requirements.txt
discord.py>=2.3.2,<3
python-dotenv>=1.0.1,<2
requests>=2.31.0,<3

